+++ src/sprite.cpp
@@ -24,7 +24,9 @@ Sprite::Sprite(Graphics &graphics, const std::string &filePath, Point source, Si
 Sprite::~Sprite() = default;
 
 void Sprite::draw(Graphics &graphics, Point point) {
-    SDL_Rect destinationRectangle = { point.x, point.y, static_cast<int>(this->sourceRect.w * globals::SPRITE_SCALE), static_cast<int>(this->sourceRect.h * globals::SPRITE_SCALE) };
+    SDL_Rect destinationRectangle = { point.x, point.y,
+  static_cast<int>(static_cast<float>(this->sourceRect.w) * globals::SPRITE_SCALE),
+  static_cast<int>(static_cast<float>(this->sourceRect.h) * globals::SPRITE_SCALE) };
 
     graphics.blitSurface(this->spriteSheet, &this->sourceRect, &destinationRectangle);
 }
